+++ src/main/ruby/truffleruby/core/enumerator.rb
@@ -231,6 +231,73 @@ class Enumerator
     end
   end
 
+  def self.product(*enums, **nil, &block)
+    return Product.new(*enums) if Primitive.nil?(block)
+    Truffle::EnumeratorOperations.product_iterator([], enums, &block)
+    nil
+  end
+
+  class Product < Enumerator
+    def initialize(*enums, **nil)
+      @enums = enums
+
+      self
+    end
+
+    def initialize_copy(product)
+      return self if Primitive.equal?(self, product)
+
+      if Primitive.class(self) != Primitive.class(product)
+        raise TypeError, 'initialize_copy should take same class object'
+      end
+
+      unless product.instance_variable_defined?(:@enums)
+        raise ArgumentError, 'uninitialized product'
+      end
+
+      # don't check explicitly whether self is frozen
+      # because assigning @enum instance variable will raise FrozenError in this case anyway
+      @enums = product.instance_variable_get(:@enums)
+
+      self
+    end
+
+    def each(&block)
+      return to_enum(:each) { size } unless block_given?
+
+      Truffle::EnumeratorOperations.product_iterator([], @enums, &block)
+
+      self
+    end
+
+    def inspect
+      return "#<#{Primitive.class(self).name}: ...>" if Truffle::ThreadOperations.detect_recursion(self) do
+        return "#<#{Primitive.class(self).name}: #{@enums || "uninitialized"}>"
+      end
+    end
+
+    def rewind
+      @enums.each do |e|
+        e.rewind if e.respond_to?(:rewind)
+      end
+
+      self
+    end
+
+    def size
+      @enums.map do |enum|
+        return nil unless enum.respond_to?(:size)
+
+        size = enum.size
+
+        return size if Primitive.is_a?(size, Float) && size.infinite?
+        return nil unless Primitive.is_a?(size, Integer)
+
+        size
+      end.reduce(1, &:*)
+    end
+  end
+
   class Yielder
     attr_accessor :memo
 
@@ -242,7 +309,6 @@ class Enumerator
 
       self
     end
-    private :initialize
 
     def yield(*args, **kwargs)
       @proc.call(*args, **kwargs)
@@ -268,7 +334,6 @@ class Enumerator
 
       self
     end
-    private :initialize
 
     def each(*args, **kwargs, &block)
       Primitive.share_special_variables(Primitive.proc_special_variables(block)) if block
@@ -309,7 +374,6 @@ class Enumerator
 
       self
     end
-    private :initialize
 
     def to_enum(method_name = :each, *method_args, **method_kwargs, &block)
       size = block_given? ? block : nil